+++ packages/ui/src/components/skip-link/skip-link.styles.ts
@@ -1,5 +1,5 @@
 import { tv } from 'tailwind-variants';
 
 export const styles = tv({
-  base: 'focus:bg-background-white-pale focus:text-text-link focus:outline-border-focus sr-only focus:not-sr-only focus:fixed focus:inset-x-0 focus:top-0 focus:p-3 focus:text-center focus:text-xl focus:underline focus:-outline-offset-2',
+  base: 'focus:focus-outline focus:bg-background-white-pale focus:text-text-link focus:focus-outline sr-only focus:not-sr-only focus:fixed focus:inset-x-0 focus:top-0 focus:p-3 focus:text-center focus:text-xl focus:underline focus:-outline-offset-2',
 });