+++ packages/ui/CHANGELOG.md
@@ -8,6 +8,7 @@
 - 95deb81: Datepicker using react-aria API
 - 6a44948: Refactored Compacta to use react-aria for improved accessibility and separated responsibilities into Compacta (container) and CompactaItem (individual item) components.
 - Button group using useToggleButtonGroup
+- focus outline of skip-link fix
 
 ## 0.50.3